requirements.txt
numpy>=1.21.0
torch>=1.12.0
scipy>=1.8.0
matplotlib>=3.5.0
pandas>=1.4.0
scikit-learn>=1.1.0
h5py>=3.6.0
tqdm>=4.64.0
pyyaml>=6.0
requests>=2.28.0
uproot>=4.3.0
awkward>=1.8.0
plotly>=5.10.0
dash>=2.6.0
hepstats>=0.6.0
pylhe>=0.3.0
particle>=0.20.0